# openapi-user.yaml
openapi: 3.0.0
info:
  title: MCCS Alpha User API
  description: |
    ## Introduction

    The MCCS Alpha User API v1 exposes all of the end user functionality currently available in the [alpha version of the MCCS API](https://github.com/technoshantoms/mccs-alpha-api).

    By providing an API, developers who want to create their own front-end user interface for MCCS will have significant flexibility to implement it in whatever way they choose. This means developers can present MCCS in any language, setup their own signup flow, optimize it for whatever devices their users prefer, develop a mobile app, integrate other services such as chat, etc.

    Importantly, an API enables developers to integrate MCCS functionality directly into their own apps (e.g., import transfer data into an accounting application, instruct mutual credit transfers from an e-wallet application, interact with MCCS data via a chat bot, etc.).

    ## Background

    To understand how the MCCS Alpha API works, please read the following design documents:

    - [MCCS Alpha Data Model](https://github.com/technoshantoms/mccs/blob/master/alpha-data-model.md)
    - [MCCS Alpha Functionality](https://github.com/technoshantoms/mccs/blob/master/alpha-functionality.md)

    ## Test Server

...

    ## License
  license:
    name: MIT License
    url: https://github.com/technoshantoms/mccs-alpha-api/blob/dev/LICENSE
  version: '1'
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /entities:
    get:
      tags:
        - Find Entities
      summary: Get a list of entities
      description: |
        Entities can be searched based on (1) their offers and wants tags, (2) the category they are listed under, (3) the date the offers and wants tags were added to their profile, (4) their entity name, and/or (5) an entity's selection of favorites.
      
        If the request does not contain a JWT along with a `querying_entity_id` specified, the search for favorites functionality will not work (e.g., all results will show the `isFavorite` flag as false).

        If a `querying_entity_id` is specified and both the requesting entity and the entity returned in the search are `tradingAccepted` status, the email address of the searched entity will also be included.
      parameters:
        - $ref: '#/components/parameters/offers'
        - $ref: '#/components/parameters/wants'
        - $ref: '#/components/parameters/category'
        - $ref: '#/components/parameters/taggedSince'
        - $ref: '#/components/parameters/entityName'
        - $ref: '#/components/parameters/favoritesOnly'
        - $ref: '#/components/parameters/queryingEntityID'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entity'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                withoutAuth:
                  value:
                    data:
                      - id: 5eec78f4a880b7c235f66e7c
                        accountNumber: "6838115832533278"
                        name: New World Pizza PLC
                        telephone: "+442098765432"
                        incType: plc
                        companyNumber: B67890
                        website: https://nwpizza.null
                        declaredTurnover: 10000
                        description: "We show you how good things can taste and where you need to go to eat them!"
                        address: "456 Yellow Brick Road"
                        city: "London"
                        region: "Greater London"
                        postalCode: "UK1 2ENG"
                        country: "England"
                        status: "pending"
                        showTagsMatchedSinceLastLogin: false
                        receiveDailyMatchNotificationEmail: false
                        offers:
                          - pizza
                          - wine
                        wants:
                          - flour
                          - mozarella
                          - tomato
                        categories:
                          - restaurant
                        isFavorite: false
                    meta:
                      numberOfResults: 1
                      totalPages: 1
                withAuth:
                  value:
                    data:
                      - id: 5eec78f4a880b7c235f66e7c
                        accountNumber: "6838115832533278"
                        name: New World Pizza PLC
                        telephone: "+442098765432"
                        incType: plc
                        companyNumber: B67890
                        website: https://nwpizza.null
                        declaredTurnover: 10000
                        description: "We show you how good things can taste and where you need to go to eat them!"
                        address: "456 Yellow Brick Road"
                        city: "London"
                        region: "Greater London"
                        postalCode: "UK1 2ENG"
                        country: "England"
                        status: "pending"
                        showTagsMatchedSinceLastLogin: false
                        receiveDailyMatchNotificationEmail: false
                        offers:
                          - pizza
                          - wine
                        wants:
                          - flour
                          - mozarella
                          - tomato
                        categories:
                          - restaurant
                        isFavorite: true
                    meta:
                      numberOfResults: 1
                      totalPages: 1
                withAuthTradingAccepted:
                  value:
                    data:
                      - id: 5eec78f4a880b7c235f66e7c
                        accountNumber: "6838115832533278"
                        name: New World Pizza PLC
                        email: nwpplc@dev.null
                        telephone: "+442098765432"
                        incType: plc
                        companyNumber: B67890
                        website: https://nwpizza.null
                        declaredTurnover: 10000
                        description: "We show you how good things can taste and where you need to go to eat them!"
                        address: "456 Yellow Brick Road"
                        city: "London"
                        region: "Greater London"
                        postalCode: "UK1 2ENG"
                        country: "England"
                        status: "pending"
                        showTagsMatchedSinceLastLogin: false
                        receiveDailyMatchNotificationEmail: false
                        offers:
                          - pizza
                          - wine
                        wants:
                          - flour
                          - mozarella
                          - tomato
                        categories:
                          - restaurant
                        isFavorite: true
                    meta:
                      numberOfResults: 1
                      totalPages: 1
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        # Allow optional Authorizations: https://github.com/OAI/OpenAPI-Specification/issues/14
        - {}
        - jwt: []
  /entities/{entityID}:
    get:
      tags:
        - Find Entities
      summary: Get a single entity
      description: Returns a single entity's details using its ID. Requests can be made without a JWT from an authenticated user (see `GET /entities` above for more information).
      parameters:
        - $ref: '#/components/parameters/entityID'
        - $ref: '#/components/parameters/queryingEntityID'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Entity'
              examples:
                withoutAuth:
                  value:
                    data: 
                      id: 5eec78f4a880b7c235f66e7c
                      accountNumber: "6838115832533278"
                      name: New World Pizza PLC
                      telephone: "+442098765432"
                      incType: plc
                      companyNumber: B67890
                      website: https://nwpizza.null
                      declaredTurnover: 10000
                      description: "We show you how good things can taste and where you need to go to eat them!"
                      address: "456 Yellow Brick Road"
                      city: "London"
                      region: "Greater London"
                      postalCode: "UK1 2ENG"
                      country: "England"
                      status: "pending"
                      showTagsMatchedSinceLastLogin: false
                      receiveDailyMatchNotificationEmail: false
                      offers:
                        - pizza
                        - wine
                      wants:
                        - flour
                        - mozarella
                        - tomato
                      categories:
                        - restaurant
                      isFavorite: false
                withAuth:
                  value:
                    data: 
                      id: 5eec78f4a880b7c235f66e7c
                      accountNumber: "6838115832533278"
                      name: New World Pizza PLC
                      telephone: "+442098765432"
                      incType: plc
                      companyNumber: B67890
                      website: https://nwpizza.null
                      declaredTurnover: 10000
                      description: "We show you how good things can taste and where you need to go to eat them!"
                      address: "456 Yellow Brick Road"
                      city: "London"
                      region: "Greater London"
                      postalCode: "UK1 2ENG"
                      country: "England"
                      status: "pending"
                      showTagsMatchedSinceLastLogin: false
                      receiveDailyMatchNotificationEmail: false
                      offers:
                        - pizza
                        - wine
                      wants:
                        - flour
                        - mozarella
                        - tomato
                      categories:
                        - restaurant
                      isFavorite: true
                withAuthTradingAccepted:
                  value:
                    data: 
                      id: 5eec78f4a880b7c235f66e7c
                      accountNumber: "6838115832533278"
                      name: New World Pizza PLC
                      email: nwpplc@dev.null
                      telephone: "+442098765432"
                      incType: plc
                      companyNumber: B67890
                      website: https://nwpizza.null
                      declaredTurnover: 10000
                      description: "We show you how good things can taste and where you need to go to eat them!"
                      address: "456 Yellow Brick Road"
                      city: "London"
                      region: "Greater London"
                      postalCode: "UK1 2ENG"
                      country: "England"
                      status: "pending"
                      showTagsMatchedSinceLastLogin: false
                      receiveDailyMatchNotificationEmail: false
                      offers:
                        - pizza
                        - wine
                      wants:
                        - flour
                        - mozarella
                        - tomato
                      categories:
                        - restaurant
                      isFavorite: true
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        # Allow optional Authorizations: https://github.com/OAI/OpenAPI-Specification/issues/14
        - {}
        - jwt: []
  /favorites:
    post:
      tags:
        - Find Entities
      summary: Create and manage a list of favorite entities
      description: A user can toggle any entity as a favorite, which makes it easy to retrieve using the `favorites_only` search parameter described above (see `GET /entities`).
      requestBody:
        $ref: '#/components/requestBodies/setFavorite'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
  /send-email:
    post:
      tags:
        - Find Entities
      summary: Send a message to an entity
      description: |
        A user can send a message by email to an entity without seeing the email address of the receiving entity. This enables the receiving entity to keep its email private until the entity operator decides to reply to the sender and continue the conversation by email.

        Entities with status `accepted`, `tradingPending`, `tradingAccepted` and `tradingRejected` can send an email to other entities with these same 4 statuses.

        Entities with status `pending` or `rejected` cannot send an email to any other entity. They also cannot receive emails because there is no way to find their ID since they do not show up in search results because of their `pending`/`rejected` status.
      requestBody:
        $ref: '#/components/requestBodies/sendEmail'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
...
components:
...
  parameters:
...
    offers:
      name: offers
      description: A list of goods/services offered by an entity
      in: query
      schema:
          type: string
      example: pizza,pasta
    wants:
      name: wants
      description: A list of good/services wanted by an entity
      in: query
      schema:
          type: string
      example: vegetables
    category:
      name: category
      description: A list of entities by category can be retrieved in the search functionality
      in: query
      schema:
        type: string
        example: restaurant
    taggedSince:
      name: tagged_since
      description: Get a list of entities that have had specified offers or wants tags added since a specific date and time
      in: query
      schema:
        type: string
        example: "2019-12-25T12:12:12.001Z"
    entityName:
      name: name
      description: A full or partial name of an entity can be searched
      in: query
      schema:
        type: string
        example: Alice's Restau
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT